+++ src/tokio.rs
@@ -3,6 +3,9 @@ pub use core::{TokioChildWrapper, TokioCommandWrap, TokioCommandWrapper};
 #[cfg(all(windows, feature = "creation-flags"))]
 #[doc(inline)]
 pub use creation_flags::CreationFlags;
+#[cfg(all(windows, feature = "job-object"))]
+#[doc(inline)]
+pub use job_object::JobObject;
 #[doc(inline)]
 pub use kill_on_drop::KillOnDrop;
 #[cfg(all(unix, feature = "process-group"))]
@@ -15,6 +18,8 @@ pub use process_session::ProcessSession;
 mod core;
 #[cfg(all(windows, feature = "creation-flags"))]
 mod creation_flags;
+#[cfg(all(windows, feature = "job-object"))]
+mod job_object;
 mod kill_on_drop;
 #[cfg(all(unix, feature = "process-group"))]
 mod process_group;